Impl user basic service, Change project structures, Fix bugs.

+++ src/main/java/com/ecar/servicestation/infra/bank/service/BankService.java
@@ -4,7 +4,8 @@ public interface BankService {
 
     String bankAccountUserAuthentication(String bankName, String accountNumber, long certificateId, String certificatePassword);
 
-    void depositTo(String bankName, String accountNumber, long amount, String msg);
+    void depositTo(String bankName, String accountNumber, int amount, String msg);
 
-    void withdrawFrom(String bankName, String accountNumber, String accessToken, long amount);
+    void withdrawFrom(String bankName, String accountNumber, String accessToken, int amount);
+
 }
